# Fix benchmark and update results