--- a/examples/benchmark.rs
+++ b/examples/benchmark.rs
@@ -1,5 +1,3 @@
-#![allow(clippy::collapsible_if)]
-
 mod utils;
 
 use std::cmp;
@@ -121,16 +119,14 @@ fn main() {
             socket.listen(1234).unwrap();
         }
 
-        if socket.can_send() {
-            if processed < AMOUNT {
-                let length = socket
-                    .send(|buffer| {
-                        let length = cmp::min(buffer.len(), AMOUNT - processed);
-                        (length, length)
-                    })
-                    .unwrap();
-                processed += length;
-            }
+        while socket.can_send() && processed < AMOUNT {
+            let length = socket
+                .send(|buffer| {
+                    let length = cmp::min(buffer.len(), AMOUNT - processed);
+                    (length, length)
+                })
+                .unwrap();
+            processed += length;
         }
 
         // tcp:1235: sink data
@@ -139,16 +135,14 @@ fn main() {
             socket.listen(1235).unwrap();
         }
 
-        if socket.can_recv() {
-            if processed < AMOUNT {
-                let length = socket
-                    .recv(|buffer| {
-                        let length = cmp::min(buffer.len(), AMOUNT - processed);
-                        (length, length)
-                    })
-                    .unwrap();
-                processed += length;
-            }
+        while socket.can_recv() && processed < AMOUNT {
+            let length = socket
+                .recv(|buffer| {
+                    let length = cmp::min(buffer.len(), AMOUNT - processed);
+                    (length, length)
+                })
+                .unwrap();
+            processed += length;
         }
 
         match iface.poll_at(timestamp, &sockets) {