Quoted from https://github.com/smoltcp-rs/smoltcp/pull/935#issuecomment-2226161931:
> Since the TCP buffer is a ring buffer, the following `recv` call may not receive all the data:
> https://github.com/smoltcp-rs/smoltcp/blob/c937695d4238c6df9556d21b464c949c7b43842c/examples/benchmark.rs#L142-L152
> 
> This is problematic because after receiving the trailing part of the ring buffer in this `poll` round, we can only receive the leading part of the ring buffer in the next `poll` round, but when we start the next `poll` round depends on many things, such as whether we have the delayed ACK timer and whether the sender sends data to the receiver immediately (the answer may be false if the receive window is less than one packet).
> 
> I believe the intention here is to receive all the data, like:
> ```rust
>         while socket.can_recv() && processed < AMOUNT {
>             let length = socket
>                 .recv(|buffer| {
>                     let length = cmp::min(buffer.len(), AMOUNT - processed);
>                     (length, length)
>                 })
>                 .unwrap();
>             processed += length;
>         }
> ```
> 
> After I made this change, the throughput of `benchmark -- --tap tap0 writer` is increased from ~1Gbps to ~8Gbps in my laptop, _without the mechanisms in this PR_.

Note that the throughput increase from 1Gbps to 8Gbps is measured on my laptop, but the diff cannot show the exact increase, either because the previous throughput statistics are outdated or because the laptop configuration has changed.